Started Changing Turret Logic.
