+++ Assets/Prefabs/EnemyPrefabs/VirusEnemy.prefab
@@ -47,6 +47,7 @@ SpriteRenderer:
   m_LightProbeUsage: 1
   m_ReflectionProbeUsage: 1
   m_RayTracingMode: 0
+  m_RayTraceProcedural: 0
   m_RenderingLayerMask: 1
   m_RendererPriority: 0
   m_Materials:
@@ -115,6 +116,9 @@ GameObject:
   - component: {fileID: 9026718276160111435}
   - component: {fileID: 5265551674960766380}
   - component: {fileID: 8876241481259301932}
+  - component: {fileID: 8504406522646984540}
+  - component: {fileID: 8943260193002671788}
+  - component: {fileID: 4572776239353991806}
   m_Layer: 0
   m_Name: VirusEnemy
   m_TagString: Enemy
@@ -160,6 +164,7 @@ MeshRenderer:
   m_LightProbeUsage: 1
   m_ReflectionProbeUsage: 1
   m_RayTracingMode: 2
+  m_RayTraceProcedural: 0
   m_RenderingLayerMask: 1
   m_RendererPriority: 0
   m_Materials:
@@ -184,6 +189,7 @@ MeshRenderer:
   m_SortingLayerID: 0
   m_SortingLayer: 0
   m_SortingOrder: 0
+  m_AdditionalVertexStreams: {fileID: 0}
 --- !u!65 &5265551674960766380
 BoxCollider:
   m_ObjectHideFlags: 0
@@ -209,8 +215,6 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 9174df1791b32f34099caeb13eb3cb94, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  nextNode: {fileID: 6823581483409553830, guid: b80b867f1978491448ef92f7c82f38a9,
-    type: 3}
   energyDrop: 5
   energyGain: 0
   health: 50
@@ -221,5 +225,45 @@ MonoBehaviour:
   spawnRate: 1
   damageValue: 10
   tilesToDisable: []
+  tilesToGoThrough: []
+  turretScript: {fileID: 0}
+--- !u!114 &8504406522646984540
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 7522207909491208702}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 47cd9bf0787b42feb05dca6d339467cd, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
   isExploding: 0
-  deactivator: {fileID: 0}
+  EMPExplosion: {fileID: 8773157126457952703, guid: b13667e3e565aa84e9da7c0c4ee5b4a5,
+    type: 3}
+--- !u!114 &8943260193002671788
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 7522207909491208702}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 40f1167581b344d2ba49187fe45b4c74, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  stateChannel: {fileID: 11400000, guid: 4943431ebaac44cf3b9f67889818544f, type: 2}
+--- !u!114 &4572776239353991806
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 7522207909491208702}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 2418e80703e8490fb1a158448966ad2f, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier:
