+++ Assets/Scripts/EnemyBehaviour/Enemy.cs
@@ -3,141 +3,59 @@ using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
 using GameGrid;
+using TurretBehaviour;
 using UnityEngine;
 
 namespace EnemyBehaviour
 {
     public class Enemy : MonoBehaviour
     {
-        private IEnumerator<GridLocation> path;
-        private Vector3 targetedPosition;
-        private float yOffset = -0.5f;
-    
-        private Motherboard motherboardScript;
-        private EnergyCounter energyCounterScript;
-        private GameManager gameManagerScript;
-        public float energyDrop;
-        public float energyGain;
-        public float health;
-        public float healthGain;
-        public float armySize;
-        public float armySizeGain;
-        public float speed;
-        public float spawnRate;
-        public float damageValue;
+        [SerializeField] private float energyDrop;
+        [SerializeField] private float energyGain;
+        [SerializeField] private float health;
+        [SerializeField] private float healthGain;
+        [SerializeField] private float armySize;
+        [SerializeField] private float armySizeGain;
+        [SerializeField] private float speed;
+        [SerializeField] private float spawnRate;
+        [SerializeField] private float damageValue;
 
-        public GameObject EMPExplosion;
-        private bool notMoving = false;
 
         public List<GameObject> tilesToDisable;
         public List<GameObject> tilesToGoThrough;
-        public bool isExploding;
         public Turret turretScript;
-    
+
+        public float Health
+        {
+            get => health;
+            set => health = value;
+        }
+
+        
+        public float HealthGain => healthGain;
+        public float EnergyGain => energyGain;
+
+        public float ArmySize => armySize;
+
+        public float ArmySizeGain => armySizeGain;
+
+        public float SpawnRate => spawnRate;
+
+        public float DamageValue => damageValue;
+
+        public float EnergyDrop => energyDrop;
+
+        public float Speed => speed;
+
         // Start is called before the first frame update
         void Start()
         {
-            energyCounterScript = GameObject.Find("Energy Counter").GetComponent<EnergyCounter>();
-            motherboardScript = GameObject.Find("Motherboard").GetComponent<Motherboard>();
-            gameManagerScript = GameObject.Find("GameManager").GetComponent<GameManager>();
+
         }
 
         public void Init(IEnumerator<GridLocation> path)
         {
-            this.path = path;
-            path.MoveNext();
-            transform.position = GridSpaceGlobalSpaceConverter.FromLocation(path.Current, yOffset);
-            AdvanceToNextNode();
+            GetComponent<EnemyPlayState>().Init(path);
         }
-
-        private bool AdvanceToNextNode()
-        {
-            if (path == null)
-            {
-                return true;
-            }
-            var isMoved = path.MoveNext();
-            if (isMoved)
-            {
-                targetedPosition = GridSpaceGlobalSpaceConverter.FromLocation(path.Current, yOffset);
-                transform.LookAt(targetedPosition);
-            }
-            return isMoved;
-        }
-
-        private bool IsReachedCurrentNode()
-        {
-            var position = transform.position;
-            return Enumerable.Range(0, 3).All(i => Math.Abs(position[i] - targetedPosition[i]) < 1e-6);
-        }
-        // Update is called once per frame
-        void Update()
-        {
-            // if we have no path, wait until we can get one.
-            if (path == null) return;
-            // transform.position = Vector3.MoveTowards(transform.position, nextNode.transform.position, speed * Time.deltaTime);
-
-            if (IsReachedCurrentNode())
-            {
-                // if there are no more nodes to go to
-                if (!AdvanceToNextNode())
-                {
-                    // destroy the object.
-                    motherboardScript.hp -= damageValue;
-                    Destroy(gameObject);
-                    return;
-                }
-            }
-        
-            transform.position = Vector3.MoveTowards(transform.position, targetedPosition, speed * Time.deltaTime);
-
-            if (health <= 0)
-            {
-                //chance for spawning a chip
-                energyCounterScript.energy += energyDrop;
-                if (isExploding)
-                {
-                    //this explodes the enemy
-                    if (!notMoving)
-                    {
-                        Vector3 spawnExplosionEffectLoco = new Vector3(transform.position.x, transform.position.y + 2, transform.position.z);
-                        Quaternion spawnExplosionRotation = new Quaternion(180, 0, 0, 180);
-                        Instantiate(EMPExplosion, spawnExplosionEffectLoco, spawnExplosionRotation);
-                        notMoving = true;
-                        StartCoroutine(disableTimer());
-                    }
-                } else
-                {
-                    Destroy(gameObject);
-                }
-            
-                return;
-            }
-        }
-
-        IEnumerator disableTimer()
-        {
-            yield return new WaitForSeconds(0.3f);
-            disableTurretsInRange(transform.position, 3.0f);
-        }
-
-        void disableTurretsInRange(Vector3 center, float radius)
-        {
-            Collider[] hitColliders = Physics.OverlapSphere(center, radius);
-
-            foreach (var hitCollider in hitColliders)
-            {
-                if (hitCollider.gameObject.TryGetComponent(out TileTurretBehavior behavior))
-                {
-                    var turret = behavior.GetTurret();
-                    if (turret != null)
-                    {
-                        turret.disableThisTurret();
-                    }
-                }
-            }
-            Destroy(gameObject);
-        }
-
     }
 }
