+++ Assets/Scripts/GameState/TestPlayState.cs
@@ -16,10 +16,6 @@ namespace GameState
 
         protected virtual void Update()
         {
-            if (Time.frameCount % 100 == 0)
-            {
-                Debug.Log("Playing Game.");
-            }
         }
 
         #region PlayPauseRegion
